Furniture dashboard (#7)
* [feat]: Furniture Dashboard - Update and Delete

With this we added the ability to see all the furniture items in the database and also added the ability to update and delete furniture items.
We also added a confirmation popup when deleting an item.
We show the 3D preview of the furniture item in the dashboard page.
The furniture items are displayed in a card format with the name, description, and image of the item.
The update furniture pages takes the old furniture data and then lets the user update the data.
The update furniture page also has a 3D preview of the item.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

Co-authored-by: Darren Victoria <darrenvictoria2@gmail.com>
Co-authored-by: Isuru Bandara <isuruushan2003@gmail.com>
Co-authored-by: Raminda Dulmin <ramiii21427@gmail.com>
Co-authored-by: Nesith Perera <nesithperera.pro@gmail.com>
Co-authored-by: Shavinka Jayamaha <shavinkajayamaha@gmail.com>

* [fix]: Furniture Preview

Fixed the issue with the furniture preview not working properly in the add furniture page.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

Co-authored-by: Darren Victoria <darrenvictoria2@gmail.com>
Co-authored-by: Isuru Bandara <isuruushan2003@gmail.com>
Co-authored-by: Raminda Dulmin <ramiii21427@gmail.com>
Co-authored-by: Nesith Perera <nesithperera.pro@gmail.com>
Co-authored-by: Shavinka Jayamaha <shavinkajayamaha@gmail.com>

---------

Co-authored-by: Darren Victoria <darrenvictoria2@gmail.com>
Co-authored-by: Isuru Bandara <isuruushan2003@gmail.com>
Co-authored-by: Raminda Dulmin <ramiii21427@gmail.com>
Co-authored-by: Nesith Perera <nesithperera.pro@gmail.com>
Co-authored-by: Shavinka Jayamaha <shavinkajayamaha@gmail.com>